Update project for Deploy button functionality
Hey @cherry! We're working on a [deploy tool](https://deploy.workers.cloudflare.com) for Cloudflare Workers apps, and we really want to include this project in the list of cool stuff to deploy!

This adds a missing piece in the deploy workflow to get it working with the tool — specifically, passing in a `CF_ACCOUNT_ID` env var based on the `CF_ACCOUNT_ID` secret, which will be set by the deploy tool.

It should pass through transparently if that secret isn't set, so it'll deploy for you as it has in the past, but for anyone that forks and provides their own secret there, it'll deploy to their account 👍